+++ src/app/components/lineage-graph/lineage-graph.component.ts
@@ -5,6 +5,7 @@ import {
     EventEmitter,
     HostListener,
     Input,
+    OnInit,
     Output,
     ViewChild,
 } from "@angular/core";
@@ -23,27 +24,26 @@ import { LineageGraphConfig, LINEAGE_CONFIG } from "./ligeage-graph.settings";
     styleUrls: ["./lineage-graph.component.scss"],
     changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class LineageGraphComponent {
+export class LineageGraphComponent implements OnInit {
     @Input() public graph: LineageGraph;
     @Input() public currentDataset: DatasetLineageBasicsFragment;
-
     @Output() public onClickNodeEvent = new EventEmitter<Node>();
-
     public readonly LineageGraphNodeKind: typeof LineageGraphNodeKind = LineageGraphNodeKind;
     public readonly DatasetKind: typeof DatasetKind = DatasetKind;
-
     public readonly LINEAGE_CONFIG: LineageGraphConfig = LINEAGE_CONFIG;
     public readonly DEFAULT_AVATAR_URL = AppValues.DEFAULT_AVATAR_URL;
-    public readonly INITIAL_GRAPH_VIEW_HEIGHT: number = screen.height - 400;
     public readonly INITIAL_GRAPH_VIEW_WIDTH: number = window.innerWidth - 120;
-
-    public view: [number, number] = [this.INITIAL_GRAPH_VIEW_WIDTH, this.INITIAL_GRAPH_VIEW_HEIGHT];
+    public view: [number, number];
     public showSidePanel = false;
 
     @ViewChild("containerRef", { static: false }) element: ElementRef;
     @HostListener("window:resize")
     public checkWindowSize(): void {
         this.changeLineageGraphView();
+    }
+
+    ngOnInit(): void {
+        this.view = [this.INITIAL_GRAPH_VIEW_WIDTH, this.lineageGraphHeight()];
     }
 
     public changeLineageGraphView(): void {
@@ -53,7 +53,7 @@ export class LineageGraphComponent {
             containerGraph.clientWidth -
             parseInt(styleElement.paddingLeft, 10) -
             parseInt(styleElement.paddingRight, 10);
-        this.view[1] = this.INITIAL_GRAPH_VIEW_HEIGHT;
+        this.view[1] = this.lineageGraphHeight();
     }
 
     public onClickNode(node: Node): void {
@@ -62,5 +62,23 @@ export class LineageGraphComponent {
 
     public onClickInfo(): void {
         this.showSidePanel = !this.showSidePanel;
+    }
+
+    private lineageGraphHeight(): number {
+        // TODO: Calculate the size of the graph taking the size of the components
+
+        // const headerViewMenuElement = document.getElementById("app-dataset-view-header");
+        // const datasetTabsElement = document.getElementById("container-dataset-tabs");
+        // const headerElement = document.getElementById("app-header");
+        // if (headerViewMenuElement && datasetTabsElement && headerElement) {
+        //     const headerHeight = headerElement.getBoundingClientRect().height;
+        //     const headerViewHeight = headerViewMenuElement.getBoundingClientRect().height;
+        //     const headerMarginTop = +getComputedStyle(headerViewMenuElement).marginTop.split("px")[0];
+        //     const datasetTabsElementHeight = datasetTabsElement.getBoundingClientRect().height;
+        //     return (
+        //         window.innerHeight - (headerHeight + headerViewHeight + headerMarginTop + 2 * datasetTabsElementHeight)
+        //     );
+        // }
+        return screen.height - 390;
     }
 }